backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
SQLAlchemy>=2.0.23
celery>=5.3.4
python-multipart>=0.0.6
pytesseract>=0.3.10
pdf2image>=1.16.3
PyPDF2>=3.0.1
opencv-python>=4.8.1
pillow>=10.0.0
PyMuPDF>=1.23.0
transformers>=4.34.1
spacy>=3.7.2
torch>=2.0.0
scikit-learn>=1.3.2
langchain>=0.0.340
pinecone-client>=2.2.4
qdrant-client>=1.6.0
sentence-transformers>=2.2.2
redis>=5.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pytest>=7.0.0
httpx>=0.25.0
